backend/requirements.txt
# Apache TacticalMesh Backend Dependencies
# SPDX-License-Identifier: Apache-2.0

# Web Framework
fastapi>=0.109.0,<0.110.0
uvicorn[standard]>=0.27.0,<0.28.0

# Database
sqlalchemy[asyncio]>=2.0.25,<2.1.0
asyncpg>=0.29.0,<0.30.0
alembic>=1.13.0,<1.14.0

# Configuration
pydantic>=2.5.0,<2.6.0
pydantic-settings>=2.1.0,<2.2.0

# Authentication
python-jose[cryptography]>=3.3.0,<3.4.0
passlib[bcrypt]>=1.7.4,<1.8.0
bcrypt>=4.1.0,<5.0.0

# Redis (optional, for caching)
redis>=5.0.0,<6.0.0

# HTTP Client (for testing)
httpx>=0.26.0,<0.27.0

# Testing
pytest>=7.4.0,<8.0.0
pytest-asyncio>=0.23.0,<0.24.0

# Development
black>=23.12.0
isort>=5.13.0
flake8>=7.0.0
mypy>=1.8.0
